pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pjj-tax-legal"
version = "1.0.0"
description = "KPMG Tax Advisory AI System"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "pydantic>=2.0.0",
    "google-generativeai",
    "google-cloud-aiplatform",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
